+++ R/session.R
@@ -20,13 +20,19 @@
 #'  `rlib_restart_package_not_found` global option to `FALSE`. In that
 #'  case, missing packages always cause an error.
 #'
-#' @param pkg The package names.
+#' @param pkg The package names. Can include version requirements,
+#'   e.g. `"pkg (>= 1.0.0)"`.
 #' @param reason Optional string indicating why is `pkg` needed.
 #'   Appears in error messages (if non-interactive) and user prompts
 #'   (if interactive).
 #' @param ... These dots must be empty.
 #' @param version Minimum versions for `pkg`. If supplied, must be the
 #'   same length as `pkg`. `NA` elements stand for any versions.
+#' @param compare A character vector of comparison operators to use
+#'   for `version`. If supplied, must be the same length as
+#'   `version`. If `NULL`, `>=` is used as default for all
+#'   elements. `NA` elements in `compare` are also set to `>=` by
+#'   default.
 #' @return `is_installed()` returns `TRUE` if _all_ package names
 #'   provided in `pkg` are installed, `FALSE`
 #'   otherwise. `check_installed()` either doesn't return or returns
@@ -54,55 +60,138 @@
 #' is_installed("utils")
 #' is_installed(c("base", "ggplot5"))
 #' is_installed(c("base", "ggplot5"), version = c(NA, "5.1.0"))
-is_installed <- function(pkg, ..., version = NULL) {
+is_installed <- function(pkg, ..., version = NULL, compare = NULL) {
   check_dots_empty0(...)
 
+  info <- pkg_version_info(pkg, version = version, compare = compare)
+  all(detect_installed(info))
+}
+detect_installed <- function(info) {
   # Internal mechanism for unit tests
   hook <- peek_option("rlang:::is_installed_hook")
   if (is_function(hook)) {
-    return(all(hook(pkg, version)))
+    return(all(hook(info$pkg, info$ver, info$cmp)))
   }
 
-  if (!all(map_lgl(pkg, function(x) is_true(requireNamespace(x, quietly = TRUE))))) {
-    return(FALSE)
-  }
-  if (is_null(version)) {
-    return(TRUE)
-  }
-
-  if (!is_character(version, n = length(pkg))) {
-    abort("`version` must be a character vector the same length as `pkg`.")
-  }
-
-  all(map2_lgl(pkg, version, function(p, v) {
-    is_na(v) || utils::packageVersion(p) >= v
+  flatten_lgl(pmap(info, function(pkg, cmp, ver) {
+    if (requireNamespace(pkg, quietly = TRUE)) {
+      is_na(ver) || exec(cmp, utils::packageVersion(pkg), ver)
+    } else {
+      FALSE
+    }
   }))
 }
+
+pkg_version_info <- function(pkg,
+                             version = NULL,
+                             compare = NULL,
+                             call = caller_env()) {
+  check_pkg_version(pkg, version, compare, call = call)
+
+  matches <- grepl(version_regex, pkg)
+  pkg_info <- as_version_info(pkg[matches], call = call)
+
+  info <- data_frame(pkg = pkg, cmp = na_chr, ver = na_chr)
+  info <- vec_assign(info, matches, pkg_info)
+
+  is_invalid <- grepl("[^A-Za-z0-9.]", info$pkg)
+  if (any(is_invalid)) {
+    elts <- encodeString(info$pkg[is_invalid], quote = '"')
+    msg <- c(
+      "Must supply valid package names.",
+      "x" = "Problematic names:",
+      set_names(elts, "*")
+    )
+    abort(msg, call = call)
+  }
+
+  if (!is_null(version)) {
+    has_version <- !detect_na(version)
+    is_redundant <- has_version & matches
+    if (any(is_redundant)) {
+      elts <- encodeString(pkg[is_redundant], quote = '"')
+      msg <- c(
+        sprintf(
+          "Can't supply version in both %s and %s.",
+          format_arg("pkg"),
+          format_arg("version")
+        ),
+        "x" = "Redundant versions:",
+        set_names(elts, "*")
+      )
+      abort(msg, call = call)
+    }
+
+    compare <- compare %||% ">="
+    compare <- compare %|% ">="
+
+    info$ver[has_version] <- version[has_version]
+    info$cmp[has_version] <- compare
+  }
+
+  if (!all(detect_na(info$cmp) | info$cmp %in% c(">", ">=", "<", "<="))) {
+    msg <- sprintf(
+      '%s must be one of ">", ">=", "<", or "<=".',
+      format_arg("compare")
+    )
+    abort(msg, call = call)
+  }
+
+  info
+}
+
+version_regex <- "(.*) \\((.*)\\)$"
+
+as_version_info <- function(pkg, call = caller_env()) {
+  if (!length(pkg)) {
+    return(data_frame(pkg = chr(), cmp = chr(), ver = chr()))
+  }
+  ver <- sub(version_regex, "\\2", pkg)
+  ver <- strsplit(ver, " ")
+
+  if (!every(ver, is_character2, n = 2, missing = FALSE, empty = FALSE)) {
+    abort(
+      sprintf("Can't parse version in %s.", format_arg("pkg")),
+      call = call
+    )
+  }
+
+  info <- set_names(transpose(ver), c("cmp", "ver"))
+  info <- map(info, flatten_chr)
+
+  pkg <- sub(version_regex, "\\1", pkg)
+  info <- c(list(pkg = pkg), info)
+
+  new_data_frame(info, .class = "tbl")
+}
+
 #' @rdname is_installed
+#' @param action An optional function taking `pkg` and `...`
+#'   arguments. It is called by `check_installed()` when the user
+#'   chooses to update outdated packages. The function is passed the
+#'   missing and outdated packages as a character vector of names.
 #' @inheritParams args_error_context
 #' @export
 check_installed <- function(pkg,
                             reason = NULL,
                             ...,
                             version = NULL,
+                            compare = NULL,
+                            action = NULL,
                             call = caller_env()) {
   check_dots_empty0(...)
+  check_action(action)
 
-  if (!is_character(pkg)) {
-    abort("`pkg` must be a package name or a vector of package names.", call = call)
-  }
+  info <- pkg_version_info(pkg, version = version, compare = compare)
+  needs_install <- !detect_installed(info)
 
-  if (is_null(version)) {
-    needs_install <- !map_lgl(pkg, is_installed)
-  } else {
-    if (!is_character(version, n = length(pkg))) {
-      abort("`version` must be a character vector the same length as `pkg`.")
-    }
-    needs_install <- !map2_lgl(pkg, version, function(p, v) is_installed(p, version = v))
-  }
+  pkg <- info$pkg
+  version <- info$ver
+  compare <- info$cmp
 
   missing_pkgs <- pkg[needs_install]
   missing_vers <- version[needs_install]
+  missing_cmps <- compare[needs_install]
 
   if (!length(missing_pkgs)) {
     return(invisible(NULL))
@@ -111,7 +200,9 @@ check_installed <- function(pkg,
   cnd <- new_error_package_not_found(
     missing_pkgs,
     missing_vers,
-    reason = reason
+    missing_cmps,
+    reason = reason,
+    call = call
   )
 
   restart <- peek_option("rlib_restart_package_not_found") %||% TRUE
@@ -119,8 +210,8 @@ check_installed <- function(pkg,
     abort("`rlib_restart_package_not_found` must be a logical value.")
   }
 
-  if (!is_interactive() || !restart) {
-    abort(cnd_header(cnd), call = call)
+  if (!is_interactive() || !restart || any(missing_cmps %in% c("<", "<="))) {
+    stop(cnd)
   }
 
   if (signal_package_not_found(cnd)) {
@@ -148,9 +239,13 @@ check_installed <- function(pkg,
   ))
 
   if (utils::menu(c("Yes", "No")) != 1) {
-    invokeRestart("abort")
+    # Pass condition in case caller sets up an `abort` restart
+    invokeRestart("abort", cnd)
   }
-  if (is_installed("pak")) {
+
+  if (!is_null(action)) {
+    action(missing_pkgs)
+  } else if (is_installed("pak")) {
     pkg_install <- env_get(ns_env("pak"), "pkg_install")
     pkg_install(missing_pkgs, ask = FALSE)
   } else {
@@ -158,25 +253,75 @@ check_installed <- function(pkg,
   }
 }
 
+check_pkg_version <- function(pkg,
+                              version,
+                              compare,
+                              call = caller_env()) {
+  if (!is_character2(pkg, missing = FALSE, empty = FALSE)) {
+    abort(
+      sprintf(
+        "%s must be a package name or a vector of package names.",
+        format_arg("pkg")
+      ),
+      call = call
+    )
+  }
+
+  if (!is_null(version) && !is_character2(version, n = length(pkg), empty = FALSE)) {
+    abort(
+      sprintf(
+        "%s must be `NULL` or a vector of versions the same length as %s.",
+        format_arg("version"),
+        format_arg("pkg")
+      ),
+      call = call
+    )
+  }
+
+  if (!is_null(compare)) {
+    if (is_null(version) || any((!detect_na(compare)) & detect_na(version))) {
+      msg <- sprintf(
+        "%s must be supplied when %s is supplied.",
+        format_arg("version"),
+        format_arg("compare")
+      )
+      abort(msg, call = call)
+    }
+  }
+}
+
+check_action <- function(action, call = caller_env()) {
+  # Take `pkg`, `version`, and `compare`?
+  if (!is_null(action)) {
+    if (!is_closure(action)) {
+      msg <- sprintf(
+        "%s must `NULL` or a function.",
+        format_arg("action")
+      )
+      abort(msg, call = call)
+    }
+    if (!"..." %in% names(formals(action))) {
+      msg <- sprintf(
+        "%s must take a %s argument.",
+        format_arg("action"),
+        format_arg("...")
+      )
+      abort(msg, call = call)
+    }
+  }
+}
+
 new_error_package_not_found <- function(pkg,
                                         version = NULL,
+                                        compare = NULL,
                                         ...,
                                         reason = NULL,
                                         class = NULL) {
-  if (!is_character(pkg)) {
-    abort("`pkg` must be character vector.")
-  }
-  if (!length(pkg)) {
-    abort("`pkg` must contain at least one package.")
-  }
-  if (!is_null(version) && !is_character(version, n = length(pkg))) {
-    abort("`version` must be a character vector as long as `pkg`.")
-  }
-
   error_cnd(
     class = c(class, "rlib_error_package_not_found"),
     pkg = pkg,
     version = version,
+    compare = compare,
     reason = reason,
     ...
   )
@@ -186,19 +331,20 @@ new_error_package_not_found <- function(pkg,
 cnd_header.rlib_error_package_not_found <- function(cnd, ...) {
   pkg <- cnd$pkg
   version <- cnd$version
+  compare <- cnd$compare
   reason <- cnd$reason
   n <- length(pkg)
 
   pkg_enum <- chr_quoted(cnd$pkg)
 
   if (!is_null(version)) {
-    pkg_enum <- map2_chr(pkg_enum, version, function(p, v) {
+    pkg_enum <- flatten_chr(pmap(list(pkg_enum, compare, version), function(p, o, v) {
       if (is_na(v)) {
         p
       } else {
-        paste0(p, " (>= ", v, ")")
+        sprintf("%s (%s %s)", p, o, v)
       }
-    })
+    }))
   }
 
   pkg_enum <- chr_enumerate(pkg_enum, final = "and")